requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
dotenv==0.9.9
fastapi==0.136.1
h11==0.16.0
idna==3.13
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
numpy==2.4.4
packaging==26.2
paho-mqtt==2.1.0
pluggy==1.6.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pymongo==4.17.0
pytest==9.0.3
pytest-mock==3.15.1
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
soupsieve==2.8.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
wheel==0.45.1
